+++ navigate/util/__init__.py
@@ -33,6 +33,7 @@ from navigate.util.dates import (
 )
 from navigate.util.naming import (
     attribute_to_setter,
+    matching_keys,
     name_contains_wildcards,
     retrieve_keys,
     wildcard_to_regex,